requirements.txt
Pillow==10.4.0
PyYAML==6.0.2
numpy==2.1.1
requests>=2.31.0
